requirements.txt
Django==5.2.9
pandas>=2.0.0
python-jobspy>=1.1.0